requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client (scraping + API calls)
httpx==0.27.2
playwright==1.49.0

# Scheduling
apscheduler==3.10.4

# Email
aiosmtplib==3.0.2
jinja2==3.1.4

# Reports — Excel
openpyxl==3.1.5

# Reports — PDF
reportlab==4.2.5

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3

# Dev / test
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.34.0
httpx==0.27.2
factory-boy==3.3.1
aiosqlite==0.20.0
